+++ Module.php
@@ -19,6 +19,8 @@ use yii\helpers\Json;
 use humhub\modules\file\libs\FileHelper;
 use humhub\libs\CURLHelper;
 use \Firebase\JWT\JWT;
+use yii\httpclient\Client;
+use yii\httpclient\Response;
 
 /**
  * File Module
@@ -117,6 +119,7 @@ class Module extends \humhub\components\Module
     {
         return $this->settings->get('verifyPeerOff');
     }
+
     public function getDemoServer()
     {
         return $this->settings->get('demoServer');
@@ -128,6 +131,27 @@ class Module extends \humhub\components\Module
                 return true;
         }
         return false;
+    }
+
+    public function getChat()
+    {
+        return $this->settings->get('chat');
+    }
+    public function getCompactHeader()
+    {
+        return $this->settings->get('compactHeader');
+    }
+    public function getFeedback()
+    {
+        return $this->settings->get('feedback');
+    }
+    public function getHelp()
+    {
+        return $this->settings->get('help');
+    }
+    public function getCompactToolbar()
+    {
+        return $this->settings->get('compactToolbar');
     }
 
     /**
@@ -213,7 +237,7 @@ class Module extends \humhub\components\Module
         return $key;
     }
 
-    public function commandService($data)
+    public function commandService($data): array
     {
         $url = $this->getInternalServerUrl() . '/coauthoring/CommandService.ashx';
 
@@ -231,24 +255,20 @@ class Module extends \humhub\components\Module
                 'body' => $data
             );
 
-            $response = $this->request($url, 'POST', $options);
-            $json = $response->getBody();
-        } catch (\Exception $ex) {
-            Yii::error('Could not get document server response! ' . $ex->getMessage());
-            return [];
-        }
+            $response = $this->request($url, 'POST', $options)->getData();
+            if (isset($response['error'])) {
+                $this->commandResponceError($response['error']);
+            }
 
-        try {
-            return Json::decode($json);
-        } catch (\yii\base\InvalidParamException $ex) {
-            Yii::error('Could not get document server response! ' . $ex->getMessage());
-            return [];
+            return $response;
+        } catch (\Exception $ex) {
+            Yii::error('CommandService: ' . $ex->getMessage());
+            return ['error' => $ex->getMessage()];
         }
     }
 
-    public function convertService($file, $ts)
+    public function fileToConversion($file, $ts, $toExt = null, $async = true)
     {
-        $url = $this->getInternalServerUrl() . '/ConvertService.ashx';
         $key = $this->generateDocumentKey($file);
 
         $user = Yii::$app->user->getIdentity();
@@ -259,14 +279,30 @@ class Module extends \humhub\components\Module
 
         $docHash = $this->generateHash($key, $userGuid);
 
-        $ext = strtolower(FileHelper::getExtension($file));
+        $fromExt = strtolower(FileHelper::getExtension($file));
+
+        $downloadUrl = Url::to(['/onlyoffice/backend/download', 'doc' => $docHash], true);
+        if (!empty($this->getStorageUrl())) {
+            $downloadUrl = $this->getStorageUrl() . Url::to(['/onlyoffice/backend/download', 'doc' => $docHash], false);
+        }
+
+        if(is_null($toExt))
+            $toExt = $this->convertsTo[$fromExt];
+
+        return $this->convertService($downloadUrl, $fromExt, $toExt, $key . $ts, $async);
+    }
+
+    public function convertService($documentUrl, $fromExt, $toExt, $key, $async = false): array
+    {
+        $url = $this->getInternalServerUrl() . '/ConvertService.ashx';
+
         $data = [
-            'async' => true,
+            'async' => $async,
             'embeddedfonts' => true,
-            'filetype' => $ext,
-            'outputtype' => $this->convertsTo[$ext],
-            'key' => $key . $ts,
-            'url' => Url::to(['/onlyoffice/backend/download', 'doc' => $docHash], true),
+            'filetype' => $fromExt,
+            'outputtype' => $toExt,
+            'key' => $key,
+            'url' => $documentUrl,
         ];
 
         try {
@@ -277,25 +313,20 @@ class Module extends \humhub\components\Module
                 $headers[$str] = 'Bearer ' . JWT::encode(['payload' => $data], $this->getJwtSecret());
             }
 
-            $options = array(
+            $options = [
                 'headers' => $headers,
                 'body' => $data
-            );
+            ];
 
-            $response = $this->request($url, 'POST', $options);
-            $json = $response->getBody();
+            $response = $this->request($url, 'POST', $options)->getData();
+            if (isset($response['error'])) {
+                $this->convertResponceError($response['error']);
+            }
+
+            return $response;
         } catch (\Exception $ex) {
-            $error = 'Could not get document server response! ' . $ex->getMessage();
-            Yii::error($error);
-            return [ 'error' => $error ];
-        }
-
-        try {
-            return Json::decode($json);
-        } catch (\yii\base\InvalidParamException $ex) {
-            $error = 'Could not get document server response! ' . $ex->getMessage();
-            Yii::error($error);
-            return [ 'error' => $error ];
+            Yii::error('ConvertService: ' . $ex->getMessage());
+            return ['error' => $ex->getMessage()];
         }
     }
 
@@ -316,14 +347,20 @@ class Module extends \humhub\components\Module
     /**
      * @inheritdoc
      */
-    public function generateHash($key, $userGuid)
+    public function generateHash($key, $userGuid, $isEmpty = false)
     {
-        $data = [
-            'key' => $key
-        ];
+        $data = [];
+
+        if (!empty($key)) {
+            $data['key'] = $key;
+        }
 
         if (!empty($userGuid)) {
             $data['userGuid'] = $userGuid;
+        }
+
+        if ($isEmpty) {
+            $data['isEmpty'] = true;
         }
 
         return JWT::encode($data, Yii::$app->settings->get('secret'));
@@ -345,37 +382,39 @@ class Module extends \humhub\components\Module
         return [$data, null];
     }
     /**
-     * @inheritdoc
+     * Send request by URL with CURL method
+     *
+     * @param string $url
+     * @param string $method
+     * @param array $options
+     * @return Response
      */
-    public function request($url, $method = 'GET', $options = [])
+    public function request($url, $method = 'GET', $options = []): Response
     {
-        $curloptions = CURLHelper::getOptions();
+        $http = new Client(['transport' => 'yii\httpclient\CurlTransport']);
+        $response = $http->createRequest()
+            ->setUrl($url)
+            ->setMethod($method)
+            ->setOptions(CURLHelper::getOptions());
+
         if (substr($url, 0, strlen("https")) === "https" && $this->getVerifyPeerOff()) {
-            $curloptions[CURLOPT_SSL_VERIFYPEER] = false;
-            $curloptions[CURLOPT_SSL_VERIFYHOST] = 0;
+            $response->addOptions([
+                CURLOPT_SSL_VERIFYPEER => false,
+                CURLOPT_SSL_VERIFYHOST => 0,
+            ]);
         }
 
-        $http = new \Zend\Http\Client($url, [
-            'adapter' => '\Zend\Http\Client\Adapter\Curl',
-            'curloptions' => $curloptions,
-            'timeout' => 10
-        ]);
-
-        $http->setMethod($method);
-
         if (array_key_exists('headers', $options)) {
-            $headers = $http->getRequest()->getHeaders();
             foreach ($options['headers'] as $nameHeader => $header) {
-                $headers->addHeaderLine($nameHeader, $header);
+                $response->addHeaders([$nameHeader => $header]);
             }
-
         }
 
         if (array_key_exists('body', $options)) {
-            $http->setRawBody(Json::encode($options['body']));
+            $response->setContent(Json::encode($options['body']));
         }
 
-        return $http->send();
+        return $response->send();
     }
 
     /**
@@ -408,4 +447,68 @@ class Module extends \humhub\components\Module
         'xlsx' => 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
         'xltx' => 'application/vnd.openxmlformats-officedocument.spreadsheetml.template'
     ];
+
+    private function convertResponceError($errorCode) {
+        $errorMessage = "";
+
+        switch ($errorCode) {
+            case -20:
+                $errorMessage = "Error encrypt signature";
+                break;
+            case -8:
+                $errorMessage = "Invalid token";
+                break;
+            case -7:
+                $errorMessage = "Error document request";
+                break;
+            case -6:
+                $errorMessage = "Error while accessing the conversion result database";
+                break;
+            case -5:
+                $errorMessage = "Incorrect password";
+                break;
+            case -4:
+                $errorMessage = "Error while downloading the document file to be converted.";
+                break;
+            case -3:
+                $errorMessage = "Conversion error";
+                break;
+            case -2:
+                $errorMessage = "Timeout conversion error";
+                break;
+            case -1:
+                $errorMessage = "Unknown error";
+                break;
+            case 0:
+                break;
+            default:
+                $errorMessage = "ErrorCode = " . $errorCode;
+                break;
+        }
+
+        throw new \Exception($errorMessage);
+    }
+
+    private function commandResponceError($errorCode) {
+        $errorMessage = "";
+
+        switch ($errorCode) {
+            case 3:
+                $errorMessage = "Internal server error.";
+                break;
+            case 5:
+                $errorMessage = "Command not correct.";
+                break;
+            case 6:
+                $errorMessage = "Invalid token.";
+                break;
+            case 0:
+                return;
+            default:
+                $errorMessage = "ErrorCode = " . $errorCode;
+                break;
+        }
+
+        throw new \Exception($errorMessage);
+    }
 }