+++ views/admin/index.php
@@ -7,7 +7,7 @@
 
 use yii\bootstrap\ActiveForm;
 use yii\helpers\Html;
-use yii\web\View;
+use yii\web\View; 
 ?>
 
 <div class="panel panel-default">
@@ -16,21 +16,17 @@ use yii\web\View;
 
     <div class="panel-body">
 
-        <?php if ($trial >= 0 && !empty($view['version'])): ?>
-            <div class="alert alert-success" role="alert">
-                <?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> successfully connected! - Installed version: {version}, trial period: {trial} days', ['version' => $view['version'], 'trial' => $trial]); ?>
-            </div>
-        <?php elseif (!empty($view['version'])): ?>
-            <div class="alert alert-success" role="alert">
-                <?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> successfully connected! - Installed version: {version}', ['version' => $view['version']]); ?>
-            </div>
-        <?php elseif (empty($model->serverUrl)): ?>
+        <?php if (empty($model->serverUrl)): ?>
             <div class="alert alert-warning" role="alert"><?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> not configured yet.'); ?></div>
-        <?php elseif (!empty($view['error']) && $view['error'] == 6): ?>
-            <div class="alert alert-danger" role="alert"><?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> invalid JWT token.'); ?></div>
-        <?php else: ?>
-            <div class="alert alert-danger" role="alert"><?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> not accessible.'); ?></div>
+        <?php elseif (!empty($error)): ?>
+            <div class="alert alert-danger error" role="alert"><?= Yii::t('OnlyofficeModule.base', 'Error when trying to connect ({error})', ['error' => $error]); ?></div>
+        <?php elseif (!empty($version) && !$trial): ?>
+            <div class="alert alert-success" role="alert"><?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> successfully connected! - Installed version: {version}', ['version' => $version]); ?></div>
+        <?php elseif (!empty($version) && $trial): ?>
+            <div class="alert alert-success" role="alert"><?= Yii::t('OnlyofficeModule.base', '<strong>ONLYOFFICE Docs</strong> successfully connected! - Installed version: {version}', ['version' => $version]); ?> <?= Yii::t('OnlyofficeModule.base', 'Trial period: {trial} days', ['trial' => $trial]); ?></div>
         <?php endif; ?>
+
+        <div class="alert alert-danger invalid-server-url" role="alert" hidden></div>
 
         <?php $form = ActiveForm::begin(['id' => 'configure-form']); ?>
         <div class="form-group">
@@ -52,9 +48,33 @@ use yii\web\View;
         </div>
 
         <div class="form-group">
+            <?= Html::activeLabel($model,'customLabel', ['class' => 'control-label']); ?>
+            <?= $form->field($model, 'chat')->checkbox(); ?>
+            <?= $form->field($model, 'compactHeader')->checkbox(); ?>
+            <?= $form->field($model, 'feedback')->checkbox(); ?>
+            <?= $form->field($model, 'help')->checkbox(); ?>
+            <?= $form->field($model, 'compactToolbar')->checkbox(); ?>
+        </div>
+
+        <div class="form-group">
             <?= Html::submitButton('Submit', ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
         </div>
 
+        <?php 
+            if(!empty($serverApiUrl)) {
+                View::registerJsFile($serverApiUrl);
+                View::registerJs('
+                    if(typeof DocsAPI === "undefined") {
+                        if ($(".error").length) {
+                            $(".error").append("<p style=\'color: #ff8989\'>' . Yii::t("OnlyofficeModule.base", "<strong>ONLYOFFICE Docs</strong> DocsAPI undefined.") . '</p>");
+                        } else {
+                            $(".invalid-server-url").html("' . Yii::t("OnlyofficeModule.base", "<strong>ONLYOFFICE Docs</strong> DocsAPI undefined.") . '");
+                            $(".invalid-server-url").show();
+                        }
+                    } 
+                ');
+            }
+        ?>
         <?php ActiveForm::end(); ?>
     </div>
 </div>