+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Added
+- editor customization
+- validation of server settings on the settings page
+
 ## 2.4.0
 ## Added
 - certificate verification setting